ftfw: show consistent information to users for problem diagnosing
This patch hides information that may confuse users while they are
diagnosing problems in their setup. For example, we hide entries
that are schedule to expire - from the user side, they are already
destroyed entries; and we show in the counters the real active entries,
not all that are stored in the caches.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>