cache_iterators: do not report ENOENT in cache_reset_timers
Do not report ENOENT to log files, this may confuse users. There's a
race condition when shortening the timers and handling the destroy
messages. However, this problem is not serious as the point of the
shortening is to reduce the lifetime of the conntracks. If the conntrack
is dying, there's no point to shorten their lifetime anymore :)

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>