ftfw: fix race condition in the helloing routine
This patch fixes a race condition that can prevent one node from sending
the initial hello message required to reset the sequence tracking.

               node A                   node B
                  |                        |
        start     |                        |
        hello msg |----------------------->|
        stop      |                        |
        start     |                        |
                  |<-----------------------| hello-back msg

In the picture above, the node A never sends the hello messages. Thus,
the node B drops the next messages as they are in the before boundary.
This patch adds a new state to the the helloing state-machine to fix
this problem.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>