doc: remove example about CacheWriteTrough
This patch removes the documentation about the CacheWriteTrhough clause.
This feature is scheduled for removal since the asynchronous nature of
conntrackd does not allow multi-path routing support. I'm lying,
actually there's a chance to support it, but we have to guarantee that
the RTT in the message synchronization between the firewall is smaller
than the RTT between the peer and the firewalls.

Moreover, this option has made more bad than good since people enable it
when things don't work. Making the whole troubleshooting more
complicated.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>