build: add attribute header size to total attribute length
This patch adds the size of the attribute header (4 bytes) to the
length field of netattr. This fixes a possible invalid memory
access in malformed messages.

This change is included in the set of scheduled changes for 0.9.9
that break backward compatibility.

This patch also removes a memset of 4096 by one to initialize
the headers and the netattr paddings.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>