ftfw: shrink alive message size
This patch reduces the size of alive messages by removing the
"from" and "to" fields which are not of any help. This patch also
removes the IS_CTL() macro since it does not return true for
the control messages anymore but only for IS_ACK(), IS_NACK() and
IS_RESYNC().

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>