cache: add status field to store the object status
This patch adds the status field to the cache object. This avoids
the (ab)use of the alarm to check if an entry is active or dead.
This is the first step to possibly move the alarm to the cache_extra
memory space of the ftfw (which is the only use by now).

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>