src: add state polling support (oppossed to current event-driven)
This patch adds the clause PollSecs that changes the normal
behaviour of conntrackd. With PollSecs set to > 0, conntrackd
polls every N seconds the entries.

This is the opposed behaviour of an event-driven behaviour but may
be useful for those that have really strong limitations in terms of
CPU consumption and want to perform a relaxed replication.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>