src: add support for approximate timeout calculation during commit
During the commit phase, the entries in the external cache entries
are inserted in the kernel conntrack table. Currently, we use a
fixed timeout that is specified in the config file. With this patch,
if you don't specify the fixed timeout value via CommitTimeout, the
daemon calculates the real timeout value during the commit phase.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>