sync-mode: flush also internal cache after reset PurgeTimeout
Currently, the daemon sends a flush request to the kernel-space.
With lots of entries and NetlinkOverrunResync disabled, the daemon
remains in an inconsistent state due to an overrun produced by the
flush report to userspace. With this patch, the daemon also flush
its internal cache after the kernel flush request.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>