conntrackd: add child process infrastructure
This patch adds a simple infrastructure that allows to account
the child processes that have been forked. This also includes
a callback handler that can be registered that is called once
the child process finishes.

We can extended this later to include an alarm to limit the
maximum lifetime of a forked child process. This is good to
ensure that child processes behave timely.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>