conntrackd: remove the cache write-through policy
This patch removes the cache write-through clause. This feature
remained undocumented although some has found it looking at the
source code. This feature has remained in the tree for quite
some time although it has several limitations. Moreover, it
is specifically broken and dangerous for Linux kernels >= 2.6.29
since it generates loops in the synchronization.

We do this removal first to prepare the introduction of a feature
to bypass the external cache.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>