conntrackd: add support to display statistics on existing child processes
This patch adds the ability to dump the list of existing child
processes. In general, it would be hard to display one since
child processes are generally forked for very specific tasks,
like commit and flush operations, and they have very limited
lifetime. However, this can be handy for debugging problems.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>