conntrackd: net message memory allocation is unsafe
We cannot assume that we will not write in the net message before we
send it, because the memory allocated for the net message (__net) is
only reserved in BUILD_NETMSG (because of the { } block in it).
This patch marks the buffer as static to avoid this problem.
Based on a patch from Samuel Gauthier <samuel.gauthier@6wind.com>

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>