conntrackd: open event handler once cache has been populated
With this patch, we open the event handler once the internal
cache (if any) is populated. This reduces the chances of a
possible premature overrun if we lauch conntrackd in a busy
firewall. However, we may still start with an internal cache
that may differ a bit from the once in the kernel.

This patch has no impact in setups where conntrackd is started
in a spare firewall.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>