conntrack: `-L --src-nat --dst-nat' filter using AND, not OR logic
The patch that I committed in 2e06d62d341fdf936dbc1fa944d5e03f761aaf0e
was incomplete. With it, `-L --src-nat --dst-nat' shows source-natted
OR destination-natted flows. This patch changes the behaviour to
show source-natted AND destination-natted flows.

This is the consistent behaviour that we expect from conntrack
(this is how it works for other options indeed).

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>