cache: log if we received a commit request while already one in progress
This patch improves the case in which we receive a commit request
but we are already performing one. This behaviour is suspicious
since the HA manager should not trigger a double master transition.
Otherwise, something probably is not configured appropriately.

This improves 98756c2608f0879a2322919c7441973216565272
"cache: close commit request if we already have one in progress".

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>