conntrack: skip sending update message to kernel if conntrack is unchanged
This speeds up operation when a lot of conntracks exist, but only
a few of them have to be altered.

This change is user-visible because the exit message
("%d flow entries have been updated") will now print the number of entries
that have been altered instead of the total number of conntracks seen.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>