conntrackd: fix checking of return value of queue_add()
Most callers of queue_add() assume that it returns != 0 in case of
success. However, it may return -1 in case that the queue gets full.
In that case, most callers have to:

- release the object that they want to enqueue.
- decrement the refcount, in case they have bumped it.

However, most of these callers are using the tx_queue which currently
has no limit in size at all. This fix is necessary in case that I
decide to limit the size of the transmission queue in the future
(which makes a lot of sense indeed).

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>