conntrackd: generalize file descriptor infrastructure
This patch generalizes the select-based file descriptor infrastructure
by allowing you to register file descriptors and its callbacks. Instead
of hardcoding the descriptors that needs to be checked.

Now, struct fds_item contains a callback and pointer to data that is
passed to it:

  struct fds_item {
         struct list_head        head;
         int                     fd;
 +       void                    (*cb)(void *data);
 +       void                    *data;
  };

Then, we check which ones are active in the select_main_step() function:

 list_for_each_entry(cur, &STATE(fds)->list, head) {
         if (FD_ISSET(cur->fd, &readfds))
                 cur->cb(cur->data);
 }

And it invoked the corresponding callback.

I had to slightly modify the channel infrastructure to fit it into
the changes.

This modularity is required for the upcoming cthelper support.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>