nfct: Update syntax to specify command before subsystem
This patch gets the nfct syntax in sync with nft so it looks like this:

	nfct <add|delete|...> object ...

instead of:

	nfct object <add|delete|...> ...

This patch retains backward compatibility so you can still use the old syntax.
The manpage and tests have been also updated to promote the adoption of this
syntax. We should have little existing clients of this tool as we can only use
this to configure the cttimeout and cthelper infrastructures.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>